Tannhauser/requirements.txt
flet==0.28.3
mysql-connector-python==9.4.0
python-dotenv==1.1.1